Add argument parsing for extended syntax, implement read/write/test command handling

+++ src/dce_info_commands.c
@@ -1,25 +1,25 @@
 #include "dce_commands.h"
 #include "dce_info_commands.h"
 
-dce_result_t dce_handle_GMI(dce_t* dce, void* group_ctx, int kind, int argc, arg_t* argv)
+dce_result_t dce_handle_GMI(dce_t* dce, void* group_ctx, int kind, size_t argc, arg_t* argv)
 {
     dce_emit_basic_result_code(dce, DCE_RC_ERROR);
     return DCE_OK;
 }
 
-dce_result_t dce_handle_GMM(dce_t* dce, void* group_ctx, int kind, int argc, arg_t* argv)
+dce_result_t dce_handle_GMM(dce_t* dce, void* group_ctx, int kind, size_t argc, arg_t* argv)
 {
     dce_emit_basic_result_code(dce, DCE_RC_ERROR);
     return DCE_OK;
 }
 
-dce_result_t dce_handle_GMR(dce_t* dce, void* group_ctx, int kind, int argc, arg_t* argv)
+dce_result_t dce_handle_GMR(dce_t* dce, void* group_ctx, int kind, size_t argc, arg_t* argv)
 {
     dce_emit_basic_result_code(dce, DCE_RC_ERROR);
     return DCE_OK;
 }
 
-dce_result_t dce_handle_GSN(dce_t* dce, void* group_ctx, int kind, int argc, arg_t* argv)
+dce_result_t dce_handle_GSN(dce_t* dce, void* group_ctx, int kind, size_t argc, arg_t* argv)
 {
     dce_emit_basic_result_code(dce, DCE_RC_ERROR);
     return DCE_OK;
